カートコンポーネントを追加
