+++ stores/cart.ts
@@ -1,5 +1,5 @@
 /* --- libs --------------------------------------------------------------------------------------------------------- */
-import { atom, RecoilState } from "recoil";
+import { atom, RecoilState, useRecoilState } from "recoil";
 import { Product } from "../types/Product";
 
 export type Cart = {
@@ -8,7 +8,6 @@ export type Cart = {
 };
 
 const initialState: Cart = {
-  // products: [{ name: "test", id: 1, price: 3 }],
   products: [],
   totalPrice: 0
 };
@@ -17,3 +16,37 @@ export const cartState: RecoilState<Cart> = atom({
   key: "cartState",
   default: initialState
 });
+
+/* --- ロジック ------------------------------------------------------------------------------------------------------- */
+export const useCart = () => {
+  const [cart, setCart] = useRecoilState(cartState);
+
+  /* --- カートへ商品追加 ----------------------------------------------------------------------------------------------- */
+  const addToCart = (product: Product): void => {
+    const newItem = cart.products.find((_product) => _product.id === product.id);
+
+    // カートに商品が入っていない場合
+    if (!newItem) {
+      product.quantity = 1;
+      setCart({
+        products: [...cart.products, product],
+        totalPrice: cart.totalPrice + product.price
+      });
+    } else {
+      // カートに商品が入っている場合
+      setCart({
+        products: cart.products.map((_product) =>
+          _product.id === newItem.id
+            ? {
+                ..._product,
+                quantity: _product.quantity + 1
+              }
+            : product
+        ),
+        totalPrice: cart.totalPrice + product.price
+      });
+    }
+  };
+
+  return { addToCart };
+};
